Create ARM_KEY_To_Increment_Decrement_HEX_Numbers
If KEY0 is pressed on the board, set the number displayed on HEX0 to 0. If KEY1 is pressed
then increment the displayed number, and if KEY2 is pressed then decrement the number. Pressing KEY3
should blank the display, and pressing any other KEY after that should return the display to 0.